// src/app/cars/models/car.ts
export interface Car {

      id: number;
      model: string;
      type: string;
      plate: string;
      deliveryDate: string;
      deadline: string;
      color: string;
      power: number;
      clientFirstName: string;
      clientSurname: string;
      cost: number;
      isFullyDamaged: boolean;
      year: number;    
      parts: Object[];
}